Attribute Macro optee_utee::ta_open_session

#[ta_open_session]

Attribute to declare the entry point of opening a session. Pointer to session context pointer (*mut *mut T) can be defined as an optional parameter.

Examples

#[ta_open_session]
fn open_session(params: &mut Parameters) -> Result<()> { }

// T is the sess_ctx struct and is required to implement default trait
#[ta_open_session]
fn open_session(params: &mut Parameters, sess_ctx: &mut T) -> Result<()> { }