Enum optee_utee::crypto_op::AlgorithmId

#[repr(u32)]
pub enum AlgorithmId {

    AesEcbNopad,
    AesCbcNopad,
    AesCtr,
    AesCts,
    AesXts,
    AesCbcMacNopad,
    AesCbcMacPkcs5,
    AesCmac,
    AesCcm,
    AesGcm,
    DesEcbNopad,
    DesCbcNopad,
    DesCbcMacNopad,
    DesCbcMacPkcs5,
    Des3EcbNopad,
    Des3CbcNopad,
    Des3CbcMacNopad,
    Des3CbcMacPkcs5,
    RsassaPkcs1V15MD5,
    RsassaPkcs1V15Sha1,
    RsassaPkcs1V15Sha224,
    RsassaPkcs1V15Sha256,
    RsassaPkcs1V15Sha384,
    RsassaPkcs1V15Sha512,
    RsassaPkcs1V15MD5Sha1,
    RsassaPkcs1PssMgf1Sha1,
    RsassaPkcs1PssMgf1Sha224,
    RsassaPkcs1PssMgf1Sha256,
    RsassaPkcs1PssMgf1Sha384,
    RsassaPkcs1PssMgf1Sha512,
    RsaesPkcs1V15,
    RsaesPkcs1OAepMgf1Sha1,
    RsaesPkcs1OAepMgf1Sha224,
    RsaesPkcs1OAepMgf1Sha256,
    RsaesPkcs1OAepMgf1Sha384,
    RsaesPkcs1OAepMgf1Sha512,
    RsaNopad,
    DSASha1,
    DSASha224,
    DSASha256,
    DhDeriveSharedSecret,
    Md5,
    Sha1,
    Sha224,
    Sha256,
    Sha384,
    Sha512,
    Md5Sha1,
    HmacMd5,
    HmacSha1,
    HmacSha224,
    HmacSha256,
    HmacSha384,
    HmacSha512,
    IllegalValue,

}

Algorithms that can be allocated as an crypto operation.

Variants

AesEcbNopad

Cipher supported algorithm.

AesCbcNopad

Cipher supported algorithm.

AesCtr

Cipher supported algorithm.

AesCts

Cipher supported algorithm.

AesXts

Cipher supported algorithm.

AesCbcMacNopad

Mac supported algorithm.

AesCbcMacPkcs5

Mac supported algorithm.

AesCmac

Mac supported algorithm.

AesCcm

AE supported algorithm.

AesGcm

AE supported algorithm.

DesEcbNopad

Cipher supported algorithm.

DesCbcNopad

Cipher supported algorithm.

DesCbcMacNopad

Mac supported algorithm.

DesCbcMacPkcs5

Mac supported algorithm.

Des3EcbNopad

Cipher supported algorithm.

Des3CbcNopad

Cipher supported algorithm.

Des3CbcMacNopad

Mac supported algorithm.

Des3CbcMacPkcs5

Mac supported algorithm.

RsassaPkcs1V15MD5

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15Sha1

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15Sha224

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15Sha256

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15Sha384

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15Sha512

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15MD5Sha1

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1PssMgf1Sha1

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1PssMgf1Sha224

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1PssMgf1Sha256

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1PssMgf1Sha384

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1PssMgf1Sha512

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsaesPkcs1V15

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaesPkcs1OAepMgf1Sha1

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaesPkcs1OAepMgf1Sha224

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaesPkcs1OAepMgf1Sha256

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaesPkcs1OAepMgf1Sha384

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaesPkcs1OAepMgf1Sha512

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaNopad

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

DSASha1

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

DSASha224

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

DSASha256

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

DhDeriveSharedSecret

DeriveKey supported algorithm.

Md5

Digest supported algorithm.

Sha1

Digest supported algorithm.

Sha224

Digest supported algorithm.

Sha256

Digest supported algorithm.

Sha384

Digest supported algorithm.

Sha512

Digest supported algorithm.

Md5Sha1

Mac supported algorithm.

HmacMd5

Mac supported algorithm.

HmacSha1

Mac supported algorithm.

HmacSha224

Mac supported algorithm.

HmacSha256

Mac supported algorithm.

HmacSha384

Mac supported algorithm.

HmacSha512

Mac supported algorithm.

IllegalValue

Reserved for GlobalPlatform compliance test applications.