Enum optee_utee::ErrorKind

#[repr(u32)]
pub enum ErrorKind {

    CorruptObject,
    CorruptObject2,
    StorageNotAvailable,
    StorageNotAvailable2,
    Generic,
    AccessDenied,
    Cancel,
    AccessConflict,
    ExcessData,
    BadFormat,
    BadParameters,
    BadState,
    ItemNotFound,
    NotImplemented,
    NotSupported,
    NoData,
    OutOfMemory,
    Busy,
    Communication,
    Security,
    ShortBuffer,
    ExternalCancel,
    Overflow,
    TargetDead,
    StorageNoSpace,
    MacInvalid,
    SignatureInvalid,
    TimeNotSet,
    TimeNeedsReset,
    Unknown,

}

A list specifying general categories of TEE error and its corresponding code in OP-TEE OS.

Variants

CorruptObject

Object corruption.

CorruptObject2

Persistent object corruption.

StorageNotAvailable

Object storage is not available.

StorageNotAvailable2

Persistent object storage is not available.

Generic

Non-specific cause.

AccessDenied

Access privileges are not sufficient.

Cancel

The operation was canceled.

AccessConflict

Concurrent accesses caused conflict.

ExcessData

Too much data for the requested operation was passed.

BadFormat

Input data was of invalid format.

BadParameters

Input parameters were invalid.

BadState

Operation is not valid in the current state.

ItemNotFound

The requested data item is not found.

NotImplemented

The requested operation should exist but is not yet implemented.

NotSupported

The requested operation is valid but is not supported in this implementation.

NoData

Expected data was missing.

OutOfMemory

System ran out of resources.

Busy

The system is busy working on something else.

Communication

Communication with a remote party failed.

Security

A security fault was detected.

ShortBuffer

The supplied buffer is too short for the generated output.

ExternalCancel

The operation has been cancelled by an external event which occurred in the REE while the function was in progress.

Overflow

Data overflow.

TargetDead

Trusted Application has panicked during the operation.

StorageNoSpace

Insufficient space is available.

MacInvalid

MAC is invalid.

SignatureInvalid

Signature is invalid.

TimeNotSet

The persistent time has not been set.

TimeNeedsReset

The persistent time has been set but may have been corrupted and SHALL no longer be trusted.

Unknown

Unknown error.