optee_utee

Module net

Structs

Setup

A setup parameter used for OP-TEE.

Socket

A struct used for socket operations.

TcpAdapter

An adapter for TCP sockets in OP-TEE. Typically, it is not used directly, but can be employed for wrapper operations, such as traffic control within the TEE.

UdpAdapter

An adapter for UDP sockets in OP-TEE. Typically, it is not used directly, but can be employed for wrapper operations, such as traffic control within the TEE.

Enums

SocketError

Traits

SocketAdapter

A trait designed to accommodate various implementations of GP TEE Sockets API.

StdCompatConnect

A trait used for convenience; import it so that the code remains consistent with the std version (with the only difference being the return error type).

StdCompatRead

A trait used for convenience; import it so that the code remains consistent with the std version (with the only difference being the return error type).

StdCompatWrite

A trait used for convenience; import it so that the code remains consistent with the std version (with the only difference being the return error type).

Type Aliases

TcpStream

A TcpStream that is compatible with OP-TEE.

UdpSocket

A UdpSocket that is compatible with OP-TEE.