optee_utee

Module object

Structs

Attribute

A general attribute (buffer or value) that can be used to populate an object or to specify operation parameters.

AttributeMemref

A buffer attribute.

AttributeValue

A value attribute.

DataFlag

A set of flags that controls the access rights and sharing permissions with which the object handle is opened.

HandleFlag

A set of flags that defines Handle features.

ObjectEnumHandle

An enumerator for PersistentObjects.

ObjectHandle

An opaque handle on an object.

ObjectInfo

Represent the characteristics of an object. This info can be returned by GenericObject function info

PersistentObject

An object identified by an Object Identifier and including a Data Stream.

TransientObject

An object containing attributes but no data stream, which is reclaimed when closed or when the TA instance is destroyed. Transient objects are used to hold a cryptographic object (key or key-pair).

UsageFlag

A set of flags that defines usages of data in TEE secure storage.

Enums

AttributeId

MiscellaneousConstants

Miscellaneous constants.

ObjectStorageConstants

TransientObjectType

Define types of TransientObject with predefined maximum sizes.

Whence

Indicate the possible start offset when moving a data position in the data stream associated with a PersistentObject.

Traits

GenericObject

A generic trait for an object (transient or persistent).