Enum optee_utee::object::TransientObjectType

#[repr(u32)]
pub enum TransientObjectType {
    Aes,
    Des,
    Des3,
    HmacMd5,
    HmacSha1,
    HmacSha224,
    HmacSha256,
    HmacSha384,
    HmacSha512,
    RsaPublicKey,
    RsaKeypair,
    DsaPublicKey,
    DsaKeypair,
    DhKeypair,
    EcdsaPublicKey,
    EcdsaKeypair,
    EcdhPublicKey,
    EcdhKeypair,
    GenericSecret,
    CorruptedObject,
    Data,
}

Define types of TransientObject with predefined maximum sizes.

Variants
Aes

128, 192, or 256 bits

Des

Always 64 bits including the parity bits. This gives an effective key size of 56 bits

Des3

128 or 192 bits including the parity bits. This gives effective key sizes of 112 or 168 bits

HmacMd5

Between 64 and 512 bits, multiple of 8 bits

HmacSha1

Between 80 and 512 bits, multiple of 8 bits

HmacSha224

Between 112 and 512 bits, multiple of 8 bits

HmacSha256

Between 192 and 1024 bits, multiple of 8 bits

HmacSha384

Between 256 and 1024 bits, multiple of 8 bits

HmacSha512

Between 256 and 1024 bits, multiple of 8 bits

RsaPublicKey

The number of bits in the modulus. 256, 512, 768, 1024, 1536 and 2048 bit keys SHALL be supported. Support for other key sizes including bigger key sizes is implementation-dependent. Minimum key size is 256 bits

RsaKeypair

Same as RsaPublicKey key size.

DsaPublicKey

Depends on Algorithm:

DsaSha1: Between 512 and 1024 bits, multiple of 64 bits

DsaSha224: 2048 bits

DsaSha256: 2048 or 3072 bits

DsaKeypair

Same as DsaPublicKey key size.

DhKeypair

From 256 to 2048 bits, multiple of 8 bits.

EcdsaPublicKey

Between 160 and 521 bits. Conditional: Available only if at least one of the ECC the curves defined in Table 6-14 with "generic" equal to "Y" is supported.

EcdsaKeypair

Between 160 and 521 bits. Conditional: Available only if at least one of the ECC curves defined in Table 6-14 with "generic" equal to "Y" is supported. SHALL be same value as for ECDSA public key size.

EcdhPublicKey

Between 160 and 521 bits. Conditional: Available only if at least one of the ECC curves defined in Table 6-14 with "generic" equal to "Y" is supported.

EcdhKeypair

Between 160 and 521 bits. Conditional: Available only if at least one of the ECC curves defined in Table 6-14 with "generic" equal to "Y" is supported. SHALL be same value as for ECDH public key size

GenericSecret

Multiple of 8 bits, up to 4096 bits. This type is intended for secret data that has been derived from a key derivation scheme.

CorruptedObject

Object is corrupted.

Data

0 – All data is in the associated data stream.