Struct optee_utee::object::ObjectEnumHandle

pub struct ObjectEnumHandle { /* private fields */ }

An enumerator for PersistentObjects.

Implementations

impl ObjectEnumHandle

pub fn allocate() -> Result<Self>

Allocate an object enumerator. Once an object enumerator has been allocated, it can be reused for multiple enumerations.

pub fn reset(&mut self)

Reset an object enumerator handle to its initial state after allocation. If an enumeration has been started, it is stopped.

pub fn start(&mut self, storage_id: u32) -> Result<()>

Start the enumeration of all the PersistentObjects in a given Trusted Storage. The object information can be retrieved by calling the function ObjectEnumHandle::get_next repeatedly.

pub fn get_next<T>( &mut self, object_info: &mut ObjectInfo, object_id: &mut [u8] ) -> Result<u32>

Get the next object in an enumeration and returns information about the object: type, size, identifier, etc.