optee_utee

Module property

Structs

ClientEndian

ClientIdentity

TaAppId

TaDataSize

TaDescription

TaDoesNotCloseHandleOnCorruptObject

TaEndian

TaInstanceKeepAlive

TaInstanceKeepCrashed

TaMultiSession

TaSingleInstance

TaStackSize

TaVersion

TeeApiVersion

TeeArithMaxBigIntSize

TeeCryptographyBsiR

TeeCryptographyBsiT

TeeCryptographyEcc

TeeCryptographyIetf

TeeCryptographyNist

TeeCryptographyOcta

TeeDescription

TeeDeviceId

TeeEventMaxSources

TeeFirmwareImplementationBinaryVersion

TeeFirmwareImplementationVersion

TeeFirmwareManufacturer

TeeInternalCoreVersion

TeeSystemTimeProtectionLevel

TeeTaPersistentTimeProtectionLevel

TeeTrustedOsImplementationBinaryVersion

TeeTrustedOsImplementationVersion

TeeTrustedOsManufacturer

TeeTrustedStorageAntiRollbackProtectionLevel

TeeTrustedStoragePersoRollbackProtection

TeeTrustedStoragePrivateRollbackProtection

TeeTrustedStorageProtectedRollbackProtection

TeeTrustedStorageRollbackDetectionProtectionLevel

Enums

PropertySet

Represents a TEE property set according to the TEE Internal API. The property set is a collection of properties that can be queried from the TEE. The property set is identified by a handle, which is a pointer to a TEE_PropSetHandle structure.

Traits

PropertyKey

Represents a TEE property key. The property key is used to identify a specific property within a property set. The property key is a string that is used to query the property value from the TEE property set. The property key is defined in the TEE Internal API, such as “gpd.client.identity” or “gpd.tee.apiversion”.

PropertyValue

Represents a TEE property value. The property value can be of different types, such as string, bool, u32, TEE_UUID, TEE_Identity, etc. The property value is obtained from the TEE property set using the TEE_GetPropertyAs* functions.