Trait optee_utee::property::PropertyKey

pub trait PropertyKey {
    type Output: PropertyValue;

    // Required methods
    fn key(&self) -> CString;
    fn set(&self) -> PropertySet;

    // Provided method
    fn get(&self) -> Result<Self::Output> { ... }
}

Represents a TEE property key. The property key is used to identify a specific property within a property set. The property key is a string that is used to query the property value from the TEE property set. The property key is defined in the TEE Internal API, such as “gpd.client.identity” or “gpd.tee.apiversion”.

Required Associated Types

type Output: PropertyValue

Required Methods

fn key(&self) -> CString

fn set(&self) -> PropertySet

Provided Methods

fn get(&self) -> Result<Self::Output>