optee_utee::property

Trait PropertyValue

pub trait PropertyValue: Sized {
    // Required method
    unsafe fn from_raw(set: TEE_PropSetHandle, key: CString) -> Result<Self>;
}

Represents a TEE property value. The property value can be of different types, such as string, bool, u32, TEE_UUID, TEE_Identity, etc. The property value is obtained from the TEE property set using the TEE_GetPropertyAs* functions.

Required Methods

unsafe fn from_raw(set: TEE_PropSetHandle, key: CString) -> Result<Self>

Safety

This function is unsafe because it dereferences raw pointers from the TEE API. The caller must ensure that the set handle is valid and that the TEE environment is properly initialized.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.