Struct optee_utee::uuid::Uuid

pub struct Uuid { /* fields omitted */ }

A Universally Unique Resource Identifier (UUID) type as defined in RFC4122. The value is used to identify a trusted application.

Implementations

impl Uuid

pub fn parse_str(input: &str) -> Result<Uuid, Error>

Parses a Uuid from a string of hexadecimal digits with optional hyphens.

Examples

let uuid = Uuid::parse_str("8abcf200-2450-11e4-abe2-0002a5d5c51b").unwrap();

pub fn from_bytes(bytes: [u8; 16]) -> Uuid

Creates a Uuid using the supplied big-endian bytes.

Examples

let bytes: [u8; 16] = [70, 235, 208, 238, 14, 109, 67, 201, 185, 13, 204, 195, 90, 145, 63, 62,];
let uuid = Uuid::from_bytes(bytes);

pub fn from_slice(b: &[u8]) -> Result<Uuid, Error>

Creates a Uuid using a slice of supplied big-endian bytes.

Examples

let bytes: &[u8; 16] = &[70, 235, 208, 238, 14, 109, 67, 201, 185, 13, 204, 195, 90, 145, 63, 62,];
let uuid = Uuid::from_slice(bytes);

pub fn new_raw(
time_low: u32,
time_mid: u16,
time_hi_and_version: u16,
clock_seq_and_nod: [u8; 8]
) -> Uuid

Crates a raw TEE client uuid object with specified parameters.

pub fn as_raw_ptr(&self) -> *const TEE_UUID

Converts a uuid to a const raw TEE_UUID pointer.