Crate optee_utee

Re-exports

pub use self::identity::Identity;

pub use self::identity::LoginType;

pub use self::arithmetical::*;

pub use self::crypto_op::*;

pub use self::extension::*;

pub use self::object::*;

pub use self::time::*;

pub use self::uuid::*;

Modules

arithmetical

crypto_op

extension

identity

net

object

property

time

trace

uuid

Macros

trace_print

Macro for printing to the trace output, without a newline.

trace_println

Macro for printing to the trace output, with a newline. Use the format! syntax to write data to the standard output. See std::fmt for more information.

Structs

Error

ParamTypes

Parameter

Parameters

TaSession

TaSessionBuilder

TeeParams

The TeeParams struct is used to manage the parameters for TEE commands.

Enums

ErrorKind

A list specifying general categories of TEE error and its corresponding code in OP-TEE OS.

ParamIndex

ParamType

Type Aliases

Result

A specialized Result type for TEE operations.

Attribute Macros

ta_close_session

Attribute to declare the entry point of closing a session. Session context raw pointer (*mut T) can be defined as an optional parameter.

ta_create

Attribute to declare the entry point of creating TA.

ta_destroy

Attribute to declare the entry point of destroying TA.

ta_invoke_command

Attribute to declare the entry point of invoking commands. Session context reference (&mut T) can be defined as an optional parameter.

ta_open_session

Attribute to declare the entry point of opening a session. Pointer to session context pointer (*mut *mut T) can be defined as an optional parameter.