Enum optee_utee::crypto_op::AlgorithmId

#[repr(u32)]
pub enum AlgorithmId {

    AesEcbNopad = 268_435_472,
    AesCbcNopad = 268_435_728,
    AesCtr = 268_435_984,
    AesCts = 268_436_240,
    AesXts = 268_436_496,
    AesCbcMacNopad = 805_306_640,
    AesCbcMacPkcs5 = 805_307_664,
    AesCmac = 805_307_920,
    AesCcm = 1_073_743_632,
    AesGcm = 1_073_743_888,
    DesEcbNopad = 268_435_473,
    DesCbcNopad = 268_435_729,
    DesCbcMacNopad = 805_306_641,
    DesCbcMacPkcs5 = 805_307_665,
    Des3EcbNopad = 268_435_475,
    Des3CbcNopad = 268_435_731,
    Des3CbcMacNopad = 805_306_643,
    Des3CbcMacPkcs5 = 805_307_667,
    RsassaPkcs1V15 = 4_026_533_936,
    RsassaPkcs1V15MD5 = 1_879_054_384,
    RsassaPkcs1V15Sha1 = 1_879_058_480,
    RsassaPkcs1V15Sha224 = 1_879_062_576,
    RsassaPkcs1V15Sha256 = 1_879_066_672,
    RsassaPkcs1V15Sha384 = 1_879_070_768,
    RsassaPkcs1V15Sha512 = 1_879_074_864,
    RsassaPkcs1V15MD5Sha1 = 1_879_111_728,
    RsassaPkcs1PssMgf1MD5 = 4_027_652_400,
    RsassaPkcs1PssMgf1Sha1 = 1_881_221_424,
    RsassaPkcs1PssMgf1Sha224 = 1_882_274_096,
    RsassaPkcs1PssMgf1Sha256 = 1_883_326_768,
    RsassaPkcs1PssMgf1Sha384 = 1_884_379_440,
    RsassaPkcs1PssMgf1Sha512 = 1_885_432_112,
    RsaesPkcs1V15 = 1_610_613_040,
    RsaesPkcs1OAepMgf1MD5 = 4_027_646_512,
    RsaesPkcs1OAepMgf1Sha1 = 1_612_775_984,
    RsaesPkcs1OAepMgf1Sha224 = 1_613_824_560,
    RsaesPkcs1OAepMgf1Sha256 = 1_614_873_136,
    RsaesPkcs1OAepMgf1Sha384 = 1_615_921_712,
    RsaesPkcs1OAepMgf1Sha512 = 1_616_970_288,
    RsaNopad = 1_610_612_784,
    DSASha1 = 1_879_056_689,
    DSASha224 = 1_879_060_785,
    DSASha256 = 1_879_064_881,
    DhDeriveSharedSecret = 2_147_483_698,
    EcDhDeriveSharedSecret = 2_147_483_714,
    EcDsaSha1 = 1_879_052_354,
    EcDsaSha224 = 1_879_056_450,
    EcDsaSha256 = 1_879_060_546,
    EcDsaSha384 = 1_879_064_642,
    EcDsaSha512 = 1_879_068_738,
    Ed25519 = 1_879_072_835,
    X25519 = 2_147_483_716,
    Md5 = 1_342_177_281,
    Sha1 = 1_342_177_282,
    Sha224 = 1_342_177_283,
    Sha256 = 1_342_177_284,
    Sha384 = 1_342_177_285,
    Sha512 = 1_342_177_286,
    Md5Sha1 = 1_342_177_295,
    HmacMd5 = 805_306_369,
    HmacSha1 = 805_306_370,
    HmacSha224 = 805_306_371,
    HmacSha256 = 805_306_372,
    HmacSha384 = 805_306_373,
    HmacSha512 = 805_306_374,
    IllegalValue = 4_026_531_839,

}

Algorithms that can be allocated as an crypto operation.

Variants

AesEcbNopad = 268_435_472

Cipher supported algorithm.

AesCbcNopad = 268_435_728

Cipher supported algorithm.

AesCtr = 268_435_984

Cipher supported algorithm.

AesCts = 268_436_240

Cipher supported algorithm.

AesXts = 268_436_496

Cipher supported algorithm.

AesCbcMacNopad = 805_306_640

Mac supported algorithm.

AesCbcMacPkcs5 = 805_307_664

Mac supported algorithm.

AesCmac = 805_307_920

Mac supported algorithm.

AesCcm = 1_073_743_632

AE supported algorithm.

AesGcm = 1_073_743_888

AE supported algorithm.

DesEcbNopad = 268_435_473

Cipher supported algorithm.

DesCbcNopad = 268_435_729

Cipher supported algorithm.

DesCbcMacNopad = 805_306_641

Mac supported algorithm.

DesCbcMacPkcs5 = 805_307_665

Mac supported algorithm.

Des3EcbNopad = 268_435_475

Cipher supported algorithm.

Des3CbcNopad = 268_435_731

Cipher supported algorithm.

Des3CbcMacNopad = 805_306_643

Mac supported algorithm.

Des3CbcMacPkcs5 = 805_307_667

Mac supported algorithm.

RsassaPkcs1V15 = 4_026_533_936

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15MD5 = 1_879_054_384

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15Sha1 = 1_879_058_480

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15Sha224 = 1_879_062_576

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15Sha256 = 1_879_066_672

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15Sha384 = 1_879_070_768

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15Sha512 = 1_879_074_864

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1V15MD5Sha1 = 1_879_111_728

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1PssMgf1MD5 = 4_027_652_400

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1PssMgf1Sha1 = 1_881_221_424

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1PssMgf1Sha224 = 1_882_274_096

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1PssMgf1Sha256 = 1_883_326_768

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1PssMgf1Sha384 = 1_884_379_440

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsassaPkcs1PssMgf1Sha512 = 1_885_432_112

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

RsaesPkcs1V15 = 1_610_613_040

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaesPkcs1OAepMgf1MD5 = 4_027_646_512

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaesPkcs1OAepMgf1Sha1 = 1_612_775_984

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaesPkcs1OAepMgf1Sha224 = 1_613_824_560

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaesPkcs1OAepMgf1Sha256 = 1_614_873_136

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaesPkcs1OAepMgf1Sha384 = 1_615_921_712

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaesPkcs1OAepMgf1Sha512 = 1_616_970_288

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

RsaNopad = 1_610_612_784

Asymmetric supported algorithm, can be applied with Encrypt or Decrypt mode.

DSASha1 = 1_879_056_689

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

DSASha224 = 1_879_060_785

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

DSASha256 = 1_879_064_881

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

DhDeriveSharedSecret = 2_147_483_698

DeriveKey supported algorithm.

EcDhDeriveSharedSecret = 2_147_483_714

DeriveKey supported algorithm.

EcDsaSha1 = 1_879_052_354

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

EcDsaSha224 = 1_879_056_450

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

EcDsaSha256 = 1_879_060_546

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

EcDsaSha384 = 1_879_064_642

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

EcDsaSha512 = 1_879_068_738

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

Ed25519 = 1_879_072_835

Asymmetric supported algorithm, can be applied with Sign or Verify mode.

X25519 = 2_147_483_716

DeriveKey supported algorithm.

Md5 = 1_342_177_281

Digest supported algorithm.

Sha1 = 1_342_177_282

Digest supported algorithm.

Sha224 = 1_342_177_283

Digest supported algorithm.

Sha256 = 1_342_177_284

Digest supported algorithm.

Sha384 = 1_342_177_285

Digest supported algorithm.

Sha512 = 1_342_177_286

Digest supported algorithm.

Md5Sha1 = 1_342_177_295

Mac supported algorithm.

HmacMd5 = 805_306_369

Mac supported algorithm.

HmacSha1 = 805_306_370

Mac supported algorithm.

HmacSha224 = 805_306_371

Mac supported algorithm.

HmacSha256 = 805_306_372

Mac supported algorithm.

HmacSha384 = 805_306_373

Mac supported algorithm.

HmacSha512 = 805_306_374

Mac supported algorithm.

IllegalValue = 4_026_531_839

Reserved for GlobalPlatform compliance test applications.