Enum optee_utee::object::TransientObjectType

#[repr(u32)]
pub enum TransientObjectType {

    Aes = 2_684_354_576,
    Des = 2_684_354_577,
    Des3 = 2_684_354_579,
    HmacMd5 = 2_684_354_561,
    HmacSha1 = 2_684_354_562,
    HmacSha224 = 2_684_354_563,
    HmacSha256 = 2_684_354_564,
    HmacSha384 = 2_684_354_565,
    HmacSha512 = 2_684_354_566,
    RsaPublicKey = 2_684_354_608,
    RsaKeypair = 2_701_131_824,
    DsaPublicKey = 2_684_354_609,
    DsaKeypair = 2_701_131_825,
    DhKeypair = 2_701_131_826,
    EcdsaPublicKey = 2_684_354_625,
    EcdsaKeypair = 2_701_131_841,
    EcdhPublicKey = 2_684_354_626,
    EcdhKeypair = 2_701_131_842,
    Ed25519PublicKey = 2_684_354_627,
    Ed25519Keypair = 2_701_131_843,
    X25519PublicKey = 2_684_354_628,
    X25519Keypair = 2_701_131_844,
    GenericSecret = 2_684_354_560,
    CorruptedObject = 2_684_354_750,
    Data = 2_684_354_751,

}

Define types of TransientObject with predefined maximum sizes.

Variants

Aes = 2_684_354_576

128, 192, or 256 bits

Des = 2_684_354_577

Always 64 bits including the parity bits. This gives an effective key size of 56 bits

Des3 = 2_684_354_579

128 or 192 bits including the parity bits. This gives effective key sizes of 112 or 168 bits

HmacMd5 = 2_684_354_561

Between 64 and 512 bits, multiple of 8 bits

HmacSha1 = 2_684_354_562

Between 80 and 512 bits, multiple of 8 bits

HmacSha224 = 2_684_354_563

Between 112 and 512 bits, multiple of 8 bits

HmacSha256 = 2_684_354_564

Between 192 and 1024 bits, multiple of 8 bits

HmacSha384 = 2_684_354_565

Between 256 and 1024 bits, multiple of 8 bits

HmacSha512 = 2_684_354_566

Between 256 and 1024 bits, multiple of 8 bits

RsaPublicKey = 2_684_354_608

The number of bits in the modulus. 256, 512, 768, 1024, 1536 and 2048-bit keys SHALL be supported. Support for other key sizes including bigger key sizes is implementation-dependent. Minimum key size is 256 bits

RsaKeypair = 2_701_131_824

Same as RsaPublicKey key size.

DsaPublicKey = 2_684_354_609

Depends on Algorithm:

DsaSha1: Between 512 and 1024 bits, multiple of 64 bits

DsaSha224: 2048 bits

DsaSha256: 2048 or 3072 bits

DsaKeypair = 2_701_131_825

Same as DsaPublicKey key size.

DhKeypair = 2_701_131_826

From 256 to 2048 bits, multiple of 8 bits.

EcdsaPublicKey = 2_684_354_625

Between 160 and 521 bits. Conditional: Available only if at least one of the ECC the curves defined in Table 6-14 with “generic” equal to “Y” is supported.

EcdsaKeypair = 2_701_131_841

Between 160 and 521 bits. Conditional: Available only if at least one of the ECC curves defined in Table 6-14 with “generic” equal to “Y” is supported. SHALL be same value as for ECDSA public key size.

EcdhPublicKey = 2_684_354_626

Between 160 and 521 bits. Conditional: Available only if at least one of the ECC curves defined in Table 6-14 with “generic” equal to “Y” is supported.

EcdhKeypair = 2_701_131_842

Between 160 and 521 bits. Conditional: Available only if at least one of the ECC curves defined in Table 6-14 with “generic” equal to “Y” is supported. SHALL be same value as for ECDH public key size

Ed25519PublicKey = 2_684_354_627

256 bits. Conditional: Available only if TEE_ECC_CURVE_25519 defined in Table 6-14 is supported.

Ed25519Keypair = 2_701_131_843

256 bits. Conditional: Available only if TEE_ECC_CURVE_25519 defined in Table 6-14 is supported.

X25519PublicKey = 2_684_354_628

256 bits. Conditional: Available only if TEE_ECC_CURVE_25519 defined in Table 6-14 is supported.

X25519Keypair = 2_701_131_844

256 bits. Conditional: Available only if TEE_ECC_CURVE_25519 defined in Table 6-14 is supported.

GenericSecret = 2_684_354_560

Multiple of 8 bits, up to 4096 bits. This type is intended for secret data that has been derived from a key derivation scheme.

CorruptedObject = 2_684_354_750

Object is corrupted.

Data = 2_684_354_751

0 – All data is in the associated data stream.